Added link to the diagram from draw.io in Drive
The downloaded file was unreadable, so I decided to share it on Google Drive. Please do not mind the way the connections look. They specify a relation of a Mandatory to a Optional object, which is not what I mean: I juist want to show a connection.  I used the ones that obeyed my wish, not the ones I would prefer to use, to save myself some time.